pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "media_retrieval"
version = "0.0.1"
description = "CLIP based project for finding images and video using natural language"
authors = [
  { name = "Tu Vu" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "hydra-core>=1.3.2",
    "ipykernel>=6.30.1",
    "pip",
    "pytest",
    "python-dotenv",
    "ruff",
    "tqdm",
    "typer",
]
requires-python = "~=3.10.0"


[tool.ruff]
line-length = 99
src = ["media_retrieval"]
include = ["pyproject.toml", "media_retrieval/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["media_retrieval"]
force-sort-within-sections = true

